Shows external.platform heirarchy more clearly.
