+++ platform/api/resources/vcluster-yaml/README.mdx
@@ -3,6 +3,8 @@ title: vcluster.yaml configuration reference
 sidebar_label: vcluster.yaml
 ---
 
+import External from '@site/platform/api/_partials/resources/config/external.mdx'
+import Platform from '@site/platform/api/_partials/resources/config/external/platform.mdx'
 import ApiKey from '@site/platform/api/_partials/resources/config/external/platform/apiKey.mdx'
 import AutoSleep from '@site/platform/api/_partials/resources/config/external/platform/autoSleep.mdx'
 import AutoDelete from '@site/platform/api/_partials/resources/config/external/platform/autoDelete.mdx'
@@ -15,7 +17,8 @@ please see [vcluster.yaml](/docs/vcluster/configure/vcluster-yaml).
 :::
 
 
-<ApiKey />
-<AutoSleep />
-<AutoDelete />
+## Config reference
+
+<External />
+
